client/requirements.txt
opencv-python
numpy
zmq
cvlib==0.2.5
Pillow==7.1.2